requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
attrs==25.3.0
bitsandbytes==0.47.0
blake3==1.0.6
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
compressed-tensors==0.10.2
cupy-cuda12x==13.6.0
cut-cross-entropy==25.1.1
DataProperty==1.1.0
depyf==0.19.0
diffusers==0.35.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dotenv==0.9.9
einops==0.8.1
email-validator==2.3.0
evaluate==0.4.6
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
fastrlock==0.8.3
filelock==3.19.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2024.6.1
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mbstrdecoder==1.1.4
mdurl==0.1.2
mistral_common==1.8.5
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
neo4j==5.28.2
networkx==3.3
ninja==1.13.0
nltk==3.9.1
numba==0.61.2
numexpr==2.12.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
openai==1.90.0
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
outlines_core==0.2.10
packaging==25.0
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
pathvalidate==3.3.1
peft==0.17.1
pillow==11.3.0
portalocker==3.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pybind11==3.0.1
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
Pygments==2.19.2
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
ray==2.49.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rouge_score==0.1.2
rpds-py==0.27.1
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
sentry-sdk==2.38.0
setproctitle==1.3.7
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sqlitedict==2.1.0
starlette==0.48.0
sympy==1.13.3
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torchao==0.13.0
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.55.4
triton==3.3.1
trl==0.22.2
typeguard==4.4.4
typepy==1.3.4
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==0.9.32
tzdata==2025.2
unsloth==2025.9.7
unsloth_zoo==2025.9.9
urllib3==2.5.0
uvicorn==0.36.0
uvloop==0.21.0
vllm==0.10.0
watchfiles==1.1.0
websockets==15.0.1
word2number==1.1
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
